+++ libs/template/jest.config.js
@@ -4,9 +4,7 @@ module.exports = {
   coverageDirectory: '../../docs/test-coverage/template',
   collectCoverageFrom: [
     './src/**/!(index).ts',
-    '!./src/lib/experimental/**/*.ts',
-    // NOTICE: ATM there is no way to test for self and globalThis in that function
-    '!./src/lib/core/utils/get-global-this.ts',
+    '!./src/lib/experimental/**/*.ts'
   ],
   snapshotSerializers: [
     'jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js',